Update BoardLogic.java
Now it's exclusively checked if new apples can be placed before a snake moves in order to give it the best possible opportunities! This also makes sure that apples are immediately spawned if there was a temporary period where no spaces were available.
Changed setApples method to optimize for this.
The game now also says when it's done!